plugins/aws-agentcore-langgraph/examples/runtime-demo/pyproject.toml
[project]
name = "runtime-demo"
version = "0.1.0"
description = "AgentCore Runtime demo - Serverless LangGraph deployment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "boto3>=1.35.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-aws>=0.2.0",
    "bedrock-agentcore",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
deploy = [
    "bedrock-agentcore-starter-toolkit",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
